finish the first example

+++ python/tilus/ir/layout/utils/cute.py
@@ -16,6 +16,7 @@ from __future__ import annotations
 
 from typing import Sequence, Union
 
+from hidet.ir.dtypes import int32
 from hidet.ir.expr import Expr
 from hidet.utils import prod
 from tilus.extensions.hidet.ir.primitives.swizzle import swizzle
@@ -107,7 +108,7 @@ class CuteSwizzle:
             # z_mask = ((1 << self.bbits) - 1) << self.mbase
             # y_mask = ((1 << self.bbits) - 1) << (self.mbase + self.sshift)
             # return offset ^ ((offset & y_mask) >> self.sshift)
-            return swizzle(offset, self.mbase, self.bbits, self.sshift)
+            return swizzle(int32(offset), self.mbase, self.bbits, self.sshift)
 
 
 class SwizzledCuteLayout:
